binance, binanceus, binancejs: unifying codebase
Creates a single base class for all binance functionality. Each individual
exchange overrides the name and urls, but otherwise uses the same functionality
